+++ docs/make.jl
@@ -14,6 +14,7 @@ makedocs(
         "Atmospheric Models" => [
             "Exponential" => "man/exponential.md",
             "Harris-Priester" => "man/harrispriester.md",
+            "Jacchia 1977" => "man/jacchia1977.md",
             "Jacchia-Roberts 1971" => "man/jr1971.md",
             "Jacchia-Bowman 2008" => "man/jb2008.md",
             "NRLMSISE-00" => "man/nrlmsise00.md"